let the playground fill the page
