+++ website/playground.md
@@ -7,12 +7,6 @@ layout: home
 import Playground from './.vitepress/components/Playground.vue'
 </script>
 
-# Playground
-
-<ClientOnly>
-	<div style="min-height: calc(100vh - 200px); margin-bottom: -90px">
-		<Suspense>
-				<Playground :is-main-playground="true" />
-		</Suspense>
-	</div>
-</ClientOnly>
+<div class="main-playground">
+	<Playground :is-main-playground="true" />
+</div>
